refactor: 사용하지 않는 모듈 삭제
