+++ src/pages/BusRoutePage/constants/busType.ts
@@ -1,0 +1,13 @@
+export const BUS_TYPE_MAP = {
+  ALL: '전체 차종',
+  EXPRESS: '대성고속',
+  SHUTTLE: '학교셔틀',
+  CITY: '시내버스',
+} as const;
+
+export const SHORT_BUS_TYPE_MAP = {
+  all: '전체',
+  express: '대성',
+  shuttle: '학교',
+  city: '시내',
+} as const;
