--- src/pages/BusRoutePage/ts/busModules.ts
@@ -1,60 +1,0 @@
-const getHour = (second: number) => Math.floor(second / 60 / 60) % 24;
-
-const getMinute = (second: number) => Math.floor(second / 60) % 60;
-
-export const getLeftTimeString = (second: number | '미운행' | undefined) => {
-  if (second === undefined) {
-    return '운행정보없음';
-  }
-
-  if (second === '미운행') {
-    return '미운행';
-  }
-
-  if (getHour(second) === 0 && getMinute(second) === 0) {
-    return '곧 도착';
-  }
-
-  if (getHour(second) === 0) {
-    return `${getMinute(second)}분 전`;
-  }
-
-  return `${getHour(second)}시간 ${getMinute(second)}분 전`;
-};
-
-export const getStartTimeString = (second: number | '미운행' | undefined, isMain: boolean = false) => {
-  if (!second) return '';
-  if (second === '미운행') return '';
-
-  const today = (new Date().getTime() - new Date().getTimezoneOffset() * 60 * 1000);
-  const arrivalTimeSecond = Math.ceil(today / 1000) + second;
-
-  const hour = getHour(arrivalTimeSecond);
-  const minute = getMinute(arrivalTimeSecond);
-
-  const timeString = [String(hour).padStart(2, '0'), String(minute).padStart(2, '0')];
-
-  if (isMain) return `${timeString[0]}시 ${timeString[1]}분`;
-  return `${timeString[0]}:${timeString[1]}`;
-};
-
-export const directionToEnglish = (direction: string) => {
-  if (direction === '한기대') return 'koreatech';
-  if (direction === '야우리') return 'terminal';
-  if (direction === '천안역') return 'station';
-  return '';
-};
-
-export const BUS_TYPE_MAP = {
-  ALL: '전체 차종',
-  EXPRESS: '대성고속',
-  SHUTTLE: '학교셔틀',
-  CITY: '시내버스',
-} as const;
-
-export const SHORT_BUS_TYPE_MAP = {
-  all: '전체',
-  express: '대성',
-  shuttle: '학교',
-  city: '시내',
-} as const;
