+++ src/pages/IndexPage/components/IndexBus/IndexBus.module.scss
@@ -13,7 +13,7 @@
 
   &__title {
     height: 53px;
-    font-family: Pretendard, sans-serif;;
+    font-family: Pretendard, sans-serif;
     font-size: 18px;
     font-weight: 500;
     color: #175c8e;
